Fix #2830: Move internal annotations to `js.annotations.internal`.
